+++ CHANGELOG.md
@@ -5,7 +5,9 @@
 - Add context to share data between different steps
 - Expose the context through environmental variables for the scripts
 - Add documentation for the environmental variables
-- Add flag to print current version
+- Add `--version` flag to print current version
+- Add `--quiet` flag to improve logging
+- Remove `-o` short flag for `--once`
 
 ## [0.2.2] - 2024-02-17